requirements.txt
boto3==1.9.71
Flask==1.0.2
Keras==2.2.4
numpy==1.15.4
Pillow==5.3.0
requests==2.21.0
https://storage.googleapis.com/tensorflow/linux/cpu/tensorflow-1.12.0-cp36-cp36m-linux_x86_64.whl
